A little reog of the ActivityInfo fields
It is annoying to have all the activities named Activity.qml
It is especially not user friendly when in QtCreator because
you have to check in which directory you are to know which
activity you are in.

This patch changes the meaning of the 'name' field of the
ActivityInfo to point to the proper qml file from
src/activities.