Fixed exit upon leaving 2 times in the same activity
We had a signal connection from the menu to the activity on the home
button. But when entered a second time another connection is done thus
the home is called twice and triggers the application quit.

In the patch we now pass the menu to ActivityBase and it triggers the
home and dialog box display.