click_on_letter: allow to initialize levels from json file
Answers and questions are loaded from a levels file to allow for
non-latin tasks. Names of levels files are expected in the form
levels-<locale>.json and must contain valid json describing the
levels (cf. the contained default levels-en.json).

Non-default levels files are supposed to live in some to-be-defined
datapath directory (for now faked). Upon loading through a File
object JSON syntax and basic payload checks are performed. If they
fail we fall back to the packaged default file: levels-en.json