translations support improved
Added a tool tools/convertPoToTsFiles/updateAll.sh that updates
all the translations/*.ts files from the *.po of a source dir
of the Gtk+ version.

Added all the locales that are well translated in the Gtk+ version.

Improved the CMake build chain to build the *.qm files in the bin
directory in one step.

Added a 'make BuildTranslations' target to create the *.qm files