core, made ApplicationInfo a singleton on the C++ side
It is not recommended to create a singleton of Qml Singleton registered
object but we could not found a better way to let us access ApplicationInfo
on the C++ side. All our test shows that it works.

This patch makes the virtual keyboard available by default on mobile
and not on desktop