core: add infrastructure for downloading and maintaining resource files
A singleton class DownloadManager has been added, that is responsible
for

  - downloading/updating Qt .rcc files for localized voices data,
  - maintaining the rcc-files locally and
  - registering the downloaded resources

It downloads from a fixed upstream http-URL to the default local
writable location. Downloads are based on common relative rcc-paths
such that a URL of the form

http://<server-base>/<path/to/my/resource.rcc>

will be downloaded to a local path

/<QStandardPaths::writableLocation(QStandardPaths::DataLocation)>/<path/to/my/resource.rcc>

and registered with a resource root path

qrc:/<path/to/my>/

Downloading and verification of local files is controlled by MD5
checksums that are expected to be stored in "Contents" files in each
upstream directory according to the syntax produced by the md5sum
tool. The checksums are used for checking whether a local rcc file is
up-to-date (to avoid unnecesary rcc downloads) and to verify that the
transfer was complete. Only valid rcc files (with correct checksums)
are registered.

The DownloadManager class has been designed flexible enough to also
account for downloading other types of rcc files (like activity-rcc
files, data rcc files, translations, etc.)

For now the class supports (at least) the following two use cases:

1. User changes locale -> upon confirmation voices rcc files for the
new locale are downloaded.

2. Program starts up -> If not prohibited by the settings check for
updates of the voices rcc file for the current locale and register it.

The added generator script (tools/generate_voices_rcc.sh) can be used
to generate voices-$LOCALE.rcc files for the upstream location.

TBD: A setting like "automaticDownloadsEnabled" should allow for
controlling whether the application may automatically download
resource updates without user confirmation. This is meant as a mean to
prohibit downloads with mobile internet connections.