core: DownloadManager: add setting downloadServerUrl to new admin group
This adds an [Admin] group to the settings supposed to hold
admin-settings, which should not be shown in the normal user config
dialog and only be editable in the config file.

The new downloadServerUrl can be used to change the upstream http
location for rcc downloads. It must point to a location with the
directory structure expected by the DownloadManager. For now only
<downloadServerUrl>data/voices/
The setting defaults to
http://gcompris.net