core: add DownloadDialog type and dialog helper functions in core.js
Popup dialogs like message boxes and friends have a dynamic nature that
does not fit well with the static QML definitions. Therefore helper
functions for showing/hiding such dialogs dynamically have been added
to core.js. Thus a dialog element does not need to be declared in QML:

1. showMessageDialog(parent, title, text, informativeText, icon,
buttonHandler) instantiates a Messagedialog object dynamically as child
of the  passed parent object.

2. showDownloadDialog(parent, properties) instantiates a DownloadDialog
object dynamically.

The DownloadDialog type visualizes download-progress of the
DownloadManager. It can be used statically (defined in QML) as well as
dynamically (through above helper in core.js). It connects to the
corresponding signals of DownloadManager and makes use of dynamic
MessageDialog objects to report success or error of the operation (if
requested).