memory, major refactoring
Now the memory engine no more analyse dataset content to find matching pair.

The caller of the memory now must provide a dataset that contains for each
level the list of images, texts and / or sounds.

Made all the math memory to be based on math table from 1 to 10.

Now the memory span the whole screen.

Improved scoring system, added an icon from openclipart to represent the
user.

Much simplified code for the memory core.