core, audio refactoring
We now have 2 global audio items, one for the voices, one for the effects.

They both can be mutted from our configuration dialog

They are both passed to the ActivityBase of the activity and they must be used
there instead of creating a GCAudio in the activity itself.

This patch does not include the refactoring of all activities.