core, major rework of the bar
Now the bar can be hidden / shown with a new bottom left button called
'bar button'.
This button cannot be changed in terms of position or size and the
bar may appear above the activity. We assume it is not an issue
because it can be hidden. The activity must take care not to put
relevant content under the bar button. On the other hand, the bar
buttons may be over relevant content even if an activity may try to
avoid it.

Now instead of specifying 'previous|next' in the content of the bar we
only specify 'level' which put the 3 buttons (previous, next and the level).

This patch adjusts all the activities to make them play nice with the new
bar.