core: VirtualKeyboard: add support for special keys and shift in particular
Special keys are modifier keys like shift/alt/meta etc. that modify
the value of a normal key. They are tracked in the key-model and allow
to change the modifier bitmask on each virtualkey.

For now only "Shift" has been implemented to allow for lower and
uppercase letters on the same VirtualKey. It is off per default and
can be activated by the shiftKey property. Note, that a client can
define the shift-modified key-value as he likes using the "shiftLabel"
property in the layout map.

Special keys are put in a seperate key row at the bottom to simplify
layout rendering.