menu, add an option to show/hide the sections
Hidding the sections can be done through a persistant option available
in the config dialog. It can also be accessed via the ctrl-s shortcut.

This option is very useful in relation to the new favorite feature.
In normal mode you set the activity you want with the favorite. Then
you hide the sections so that the children see only the favorite. When
the section is hidden you cannot change the favorite.