wordlist, replace full random by a shuffled list
Getting the next word randomly makes it possible to ask several times
the same word and miss other one in the list.

With this change we now create a shuffled list of the words in a level.
Then we return the next item in this queue on each getRandomWord() call.

To create the initial list you must call initRandomWord(level).
When a child misses a word, call appendRandomWord(word) to put is back
in the queue.