core, locale support for the 'system' value
We were saving the current locale in the config file. It was confusing
becasuse it was not really the user choice and for example makes it
impossible to support a different locale if the user system has changed
or if he uses the LANG environmental variable.

Now in the configuration we save the value 'system' by default which
means no choice was done by the user and that we must use the system
one. Added ApplicationInfo::getLocaleShort(const QString &locale) to
make it easy to manage in the configuration dialog.

In the config now the 'your system default' line is selected if the
config is 'system'.