core, DialogConfig: allow to set font-size base
For all font-sizes set via GCText.fontSize we add a baseFontSize
taken from settings.

Through this base value we allow the user to set the overall font-size
base in the config dialog.

For now we update this base summand in DialogConfig onValueChange,
i.e. let the changes take effect immediately. A minor improvement
could be to reset these changes upon Escaping the DialogConfig.