core bonus, no more need audioEffects to be passed, play voices
Now Bonus sound works by default without having to be passed audioEffects
which is taken in the global scope.

Add support to play the voices as it was done in the Gtk+ version. It
the voice is not found, fallback on the audio effect.

We play the voice through the audioEffects in order to have it in
sync with the bonus.