core, add support on Android to put the voices files in the apk
It may be desirable to ship an apk that contains all the resources we
have, especially when it is to be deployed on systems without Internet.

To this end, we now check the presence of the resources in the asset
directory. To put them, before doing the apk you nust put the files
under android/assets in your build directory, like:
android/assets/data/voices/voices-fr.rcc
android/assets/data/voices/Contents
android/assets/data/words/words.rcc
android/assets/data/words/Contents