core: fix locale handling for voices
For voices we need to be more specific with locale-names as we provide
some voices-packages that are territory/country-specific. So far we
used pt for pt_BR and could not resolve zh_CN correctly.

Therefore we added a new helper, ApplicationInfo::getVoicesLocale()
that returns a correct voices locale name and expects as input a
locale of the form <language>_<country>.

This is used then for voices-downloads as well as for
getAudioFilePath().