core, add an 'enabled' property at ActivityInfo level
This property can be used to hide an activity if it does not meet some
criteria. The main use case is to hide activity that requires Internet
when Internet is not allowed.

Use this feature in imageId in case there is no internet and we don't already
have the images.