main, regression, don't re-start() activity after leaving dialog
This was introduced by changes to stackView interaction for
balancebox.

In stackView transitions we now honour an alwaysStart property that
activities/dialogs can set to true to always enforce emission of the
start() signal when becoming visible. So far this was only triggered
if (!exitItem.isDialog).

This is needed for special cases in view transitions (like config
dialog to editor in balancebox) when the entering item always needs to
be start()ed.