Build: rework qml-box2d integration
We now ...

... automatically try to use a system-installed qml-box2d plugin and
    fall back to submodule build (-DQML_BOX2D_MODULE=auto, default)
... can enforce to only check for a system-installed module and error
    out if we don't find such a module (-DQML_BOX2D_MODULE=system)
... can enforce a build from the submodule
    (-DQML_BOX2D_MODULE=submodule)
    In this case we also automatically try to init && update the
    submodule on the fly.

A module from the submodule build is always installed. A system
installed module is only installed, i.e. packaged for platforms, that
need it: So far Android+SailfishOS, to-be-completed ...

Note: Also updated the qt_helper.cmake as not all Qt versions seem to
ship a Qt5::QMinimalIntegrationPlugin cmake module.