core, add loading overlay and activate it while loading activities
The loading overlay is meant to signal heavy work to the user, and
should be used whenever the ui feels like  frozen. Therefore it
deactivates all ui-interaction (except key-handling).

There should be only one instance of Loading in the whole application,
which is a sibling of the stack-view. Therefore we pass it into the
activities via the ActivityBase.loading property, as we do e.g. with
audioEffects.

For now we activate it always during activity loading through the
activity-loader. Especially on (slower) Android devices the intial
rendering of the activity-scene takes some time, and so far looked
like a freezing ui to the user.

We might want to activate the loading overlay only on mobile devices,
as a desktop should render its scene pretty fast.