balancebox, don't freeze the ui during map creation
Use the loading overlay to deactivate user-interaction during the
expesive rendering of the box-scene in the playing and the editor
view. In both cases we need to rework the expensive scene changes or
the blocking of the ui will also affect the loading animation, which
then will not even appear.

In the editor we use a WorkerScript to do the ListModel changes in
another thread. Works smoothly on Linux. On android seems that this
does actually not happen in another thread, but asynchronously on the
ui thread.

For the playing map we now instantiate our dynamic objects now
asynchronously (via incubateObject()), which takes longer than before,
but works smoothly on all tested platforms.