activation, add an internal activation code (no inapp)
Replaced CMake WITH_ACTIVATION_CODE by ACTIVATION_MODE. It can take
3 values, none, inapp or internal.
Add support for entering an activation code in the config dialog.
Add a new persistent config value named 'key' that holds the activation code.
On startup the key is checked in internal mode the key is checked and the
demo mode is set if the code does not match.
The configuration includes an entry box to enter the activation code.
Created another BuyMeOverlay dialog for to have one for inapp and one for
internal mode.