tangram, added an edition mode
Set the edition mode by changing items.editionMode. It is not accessible
by the user. The edition mode add a relod button in the bar. This button
save the level in /tmp. Then the level can be copyed in dataset.js

The file toBeAligned.js contains a lot of levels that have to be aligned
manually.