core, replace and improve command line disable_config
The disable-config is in fact coded as a disable-kioskmode. This patch changes
the command line switches.

Also the disabling of the kiosk mode was not removing the quit button and the esc
button which makes sense to remove for a kiosk mode.

I did left ctrl-q to quit GCompris supposing the children won't find it.