categorization:expert mode only on some categories
The expert mode on the categorization activity chooses randomly
from all the available categories. But among them, there may be
few categories which doesn't need an expert mode (example: the odd
even category)

The indices of the categories which are loaded in the current level
and which contains the expert mode is stored in an array. While
choosing a random index during expert mode, it is checked if it is
present in the array, if not, the randomisation process is repeated

Signed-off by: Rudra Nil Basu <rudra.nil.basu.1996@gmail.com>