expert mode:dataset selection independent of index
Selecting the categories for the expert mode is now independent
of the index of the level. "expertCategories" directly stores the
content of the categories which supports expert mode. While randomly
selecting categories for the expert mode, a random category is
selected from the "expertCategories" array directly

Signed-off by: Rudra Nil Basu <rudra.nil.basu.1996@gmail.com>