lang, fix character encoding issues in content-ml.json
fix all character encoding issues in content-ml.json
It appeared that many words when typed in (kate) text editor had different encoding
than the same word typed in the activity, resulting in answers not validated
because the strings encoding don't match,
although both the question and answer display exactly the same thing in the activity.

The method to fix this was to log to console the answers typed in the activity,
then copy them from the console and paste again in (kate) text editor.

Result was fully tested using both Raghumalayalam Sans and Noto Sans Malayalam fonts,
all work as expected.