core, fix resizing of Activity Settings page
fix the layout of Activity Settings page to always adapt to window resizing,
and also clean a little the code

ToFix: there is still a bug sometimes when resizing a lot and very fast the window,
that makes all the application disappear (empty white window).
I could only reproduce this on Activity Settings page when opened from the menu,
not when opened from within an activity.
Bug was already there before my changes, and doesn't seem directly related to that page content
(I could even reproduce it with all the visual content removed except the empty purple background)