enumerate, hide instruction on item drag
automatically hide the instruction when starting to drag an item,
and show it again when loading next level with new text.

I feel it's better this way as the instruction can hide items,
and also can overlap the answer area,
while it is not really necessary to keep it displayed to be able to complete a level.