explore_farm_animals, fix playing audio after bonus
fix playing the audio after the bonus

The audio for second level was cut by the bonus voice.
This makes sure it is played only after the bonus voice is finished.
Also in case there is no voice, a sound is played with GCSfx instead,
so adding similar done signal for it.