graph-coloring, fix keyboard focus issue after config change
there was a bug that after using the combobox inside the activity config,
the keyboard focus was lost when coming back in the activity.

Using forceActiveFocus() on the background worked, but then keyboard events were duplicated,
unless using event.accepted = true... but then escape didn't work anymore to close the help
popup or the activity...

So the best solution I found is to handle the input in a dedicated item,
and use forceActiveFocus on it.