chess, layout refactoring
improve and simplify the layout
make sure that the buttons are always visible,
and that nothing is overlapping.

There is still a not-optimal corner-case if the screen is horizontal but almost square
AND the bar is hidden, but in that case user should just not hide the bar...
I believe it works good for all practical use-cases like this.