gnumch, fix animation bug
fix muncher occasional animation bug

there was a bug sometimes, apparently depending on the screen ratio,
where the muncher AnimatedSprite was playing currentFrame 4 to 7 while having only 4 frames (!),
which made the animation loop forever as the condition to stop it was never met.

The fix is to add an extra independent variable to have a proper counter.

Also, simplified a bit the code and spritesheet images to always use square
frames (80px height and width), which also helps to get cleaner rendering.

Preparing the field for new drawings that will come later...