analog_electricity, update images for components and fix 2 issues with netlist
new images for all remaining components
improve components size calculation

Also fix two simulation issues:
-Fix errors caused by 0.001 resistance when it should be 0 (in switches and rheostat)
so we changed back those values to 0, and add the items to the vSourcesList in that case
as the engine will return one more current result in the list).
-Fix issues caused by 100000000 resistance (in switches, bulb and redLed)
this value was used to simulate disconnections internally like in the old GCompris,
but gave incorrect voltage results in some cases,
so instead we just don't push the item to the netlist if broken or switched off.