balancebox, move config to activityconfig (multiple dataset style)
balancebox, move config to activityconfig (multiple dataset style)
As the editor and loading of user level files is dependant on having the activity launched,
we only enable those buttons when the activity settings menu is started from the activity.
When started from the menu, we disable the buttons and change their text accordingly.