number_sequence, improve usablitity
change the points size to be always proportional to the image

double the size of current point to make it more usable on small touchscreen,
and it also improves the usablility in draw letters and draw numbers to have a bit more tolerance

note: for number_sequence itself, now the current point to touch is always bigger
which gives a visual hint.
The problem is that the levels were anyway not very good with some points overlapping a lot
in some places, which make it not really possible to actually count each point.
We would need to entirely redo new levels for number_sequence to avoid any overlap
and have all numbers always as big as current point.