core, add clickable-link support in DialogBackground
to be able to open the link in a browser when clicking it,
the text item needs this onLinkActivated line.

note: it looks like it only works when the link is directly in the text,
and not when it is passed as argument.
So, we should probably avoid passing links as arguments in the future.