babymatch, fix sourcesize for svg images
set sourcesize for all images to improve svg rendering
note that since the original sourcesize of png images is used in some levels
of certain activities, it was needed to use some extra duplicate Image items (not visible),
just to store the original images to fetch their size when needed.

Another side note: I noticed that even with sourcesize set, the images in DragListItem
are not perfectly rendered, a bit blurred. I'm still not sure why it happens...
(tried to round the size and position, made sure that size/sourcesize was the same,
to set the values in several ways... the same image with exact same size and sourcesize always
looks better when its parent is the backgroundImage than when it's the list tile... weird)