geo-country, update map of Norway to new county structure
In January 2020, several of the counties of Norway were merged,
reducing the number of counties down from 19 to 11. This commit
updates the map and is based on the official map from the
Norwegian Mapping Authority (Kartverket), available at
https://www.kartverket.no/til-lands/kart/illustrasjonskart
and released under the CC BY 4.0 license (with ‘© Kartverket’
as the attribution text).