geography, update continents map to svg
Start the work to update geography maps to svg.

This is just the beginning, some code need to be changed to improve
the rendering of the background map,
and I need to find a way to automate the export of svg items and the
calculation of their center coordinates to make this update and
future map creation easier.