tools, add script to convert svg to plain svgz
this is needed especially for geography maps, as they can be quite heavy
so converting them to plain svgz is saving lot of space.

For other svg images it may not be relevant to compress them all to svgz, as it would
make the whole application more computationally heavy in general for small size benefits.