reversecount, don't move if number is smaller than target number
before, tux would move if the domino's number is smaller than the target,
but not move if it's bigger than the target number.

Now it will move only if the domino's number is equal to the target number,
and if not there is a sound feedback with the clock position decrease.