reversecount, better way to check if tux can move
use a specific bool to check if tux can move, which is set to false when he
starts moving and then true when the fish opacity goes back from 0 to 1.

It fixes the issue that ok could be clicked as soon as he started to move to
last block, which was too early.

Ideally using the existing bool tuxIsMoving would have been good, though as
it's already used heavily for the logic to move from block to block it would
need quite a refactoring, so better use an extra specific bool for this.

Also not totally related change, in moveTux() function from the js file,
refactor the conditions to first check if domino's value is 0,
then if the value is wrong, then else (so if it's not 0 and right answer).