guesscount, fix result not loaded in second row if first level contains more than 1 operator
if there was more than 1 operator in first level (when using the Admin mode,
or changing the default dataset), the result of the first row was not loaded
at the beginning of the second row.

it happened because for some reason the repeater, on it's first load, was
creating instances in reverse order from the last to the first, hence breaking
the binding logic in the delegate.

the workaround here was to use an intermediate array.