blancebox, new images
new images for balancebox

images have 1px empty border to avoid bad aliasing during tilt transformations.

add missing sourceSizes for better scaling.

use 1.1 scale on the Goal image to avoid empty border when it's near a wall,
as it looks better.

add a transparent whit overlay on the background to better focus on the central
map.