core, improve filtering of character for on-screen keyboard
The previous filtering of characters to display for the on-screen
keyboard was both too strict and not strict enough. It disallowed
normal hyphens (used in, e.g., the English title of the ‘Control
the hose-pipe’ activity), but allowed soft hyphens, which are
normally invisible. Now, only letters, digits and the normal hyphen
are allowed.