comparator, little changes in datasets 8, 9 and 10
in dataset 8, 9 and 10, use a smaller maxDistanceBetweenNumbers for each of
their first level, and increase it up to 1 for each of their last level.

This way each of these datasets start with a lower probability to have different
integer parts in a pair, and this probability goes around 50% for their
last level.