menu, fix possible overflow of categories text on categories icon
in GCButton, change haveIconRight bool (which was actually not used) to
a rightIconSize real which gives the right margin value for the text if there's
an icon on the right.

also add a little margin on all sides of the text in GCButton to avoid collision
with rectangle's border and improve readability.