core, improve logo position in DialogAbout.qml
move the logo to the top, and center it

Note: it looks like the HTML subset can only center an image horizontally if
there's text on the same line. So as a workaround, I've added one full stop
character on each side, with a transparent color, and it does the trick.