core, activities, fix Tutorial focus issues and simplify it
fix focus issues when opening a Dialog (help or config) while Tutorial
is visible, by replacing lots of focus rules with some simple Keys.forwardTo
in all activities using it.

Also in planegame, use directly a Tutorial with conditions to show it instead
of using a Loader, to make it easier to use Keys.forwardTo on it.