core, replace Home button with Quit button when launching a standalone activity
Replace Home with Quit button when using a command-line to launch directly
an activity. So it is not needed to pass through the menu to quit, and also
makes it properly a standalone activity launcher.

Note that the button is not replaced when using the kiosk mode, as it would
break the "no quit button" aspect of the kiosk mode, and also removing the
home button in kiosk mode would make it impossible to go back in activities
with a sub-menu like lang or solar_system.