core, update download rcc way
Each rcc file is now suffixed with the date of its last modification and will be stored in data3/ folder.
Each resource has a type, either words, voices, backgroundMusic or full so we do't hardcode too much the paths to resources.

Handle the full rcc as other rcc files, it's no more handled separately in the ActivityInfoTree constructor.