renewable_energy, apply qmllint
Looks like there is a bug with Qt6 new PropertyChange syntax replacing "target:".
When using it in a State, then NumberAnimation and PropertyAction in a Transition
don't automatically get the "to:" and "value:" from target state anymore.

Best workaround to avoid repeating state values is to use in NumberAnimation
and PropertyAction: property: "targetID.propertyName"