menu, fix slower category selection
looks like with Qt6, category selection in the menu is 2 to 4 times slower than before.
After investigation, I noticed that most of the time was spent
when Q_EMIT menuTreeChanged() was called.

This commit makes sure that it is not called more than once when it's needed,
and it seems to remove this extra slowness.