core, make --difficulty option work when combined with --launch option
Before when using both options, it was still launching the activity with
difficulty values set in the config; but the dataset panel showed the datasets
corresponding to the difficulty range set in the command. That was not consistent
and potentially confusing.

Now both options work nicely together; and we make sure that the launched activity
has some levels corresponding to the requested difficulty range, else we stop
(same as when giving an invalid difficulty range value).