core, disable wheel scrolling of config pages when combobox popup is open
Looks like it's not directly possible to accept wheel events on a Flickable
(or at least not without overriding it with a custom MouseArea or WheelHandler)
So best solution is to set its interactive value to false when another
Flickable is on top.

In this case we can bind to the page focus value as we already handle switching
that for keyboard scrolling.