click_on_letter, better way to display question in case letter audio is not found
Check for each letter if the audio is found, else show the letter.
For example currently with Brezhoneg language, there are voices but some letters
from default dataset are missing.

NOTE: warn translators to double-check that either they provide a levels json
file for their language (with of course only letters we have audio for in their
language), or make sure that we have audio in their language for all letters from
the default levels-en.json file.