core, fix occasional focus issues caused by DialogChooseLevel
Fix a bug I noticed at least in enumerate, when starting it directly
with the --launch option, then opening the DialogChooseLevel and closing it
with "Save", the activity focus was lost.

I tried all the ways to force it back, which didn't work at all, because it
was actually passing the activeFocus to the activity once, but then taking it
again and then passing it to the StackView...

So the cleanest way I found to avoid the issue is to switch the dialog focus off
before closing it.