server, fix displaying pupils groups in ManagePupilsView list
Fix the split in DatabaseController::selectToJson to properly display pupils
groups in ManagePupilsView's "Pupils and groups" list (the previous split was not
reliable, this one checks for COMMA_SEPARATOR but not COMMA_CRYPTED_SEPARATOR).

Also add a space after the comma for display readability.

Thanks to Aish' for the initial RegExp help!