menu, improve status indicator for teacher's server
Add a bar button to show server status, but only displayed after first
connection to teacher's server (so it's not visible when not used with the
server, but stays visible for all the session once it has been connected...).

Also add a dialog with a message telling the server status when clicking on
that bar button (so it's more obvious which color corresponds to which state).