vertical_subtraction, add text warning for negative number results
Add a clear text warning in case the result is a negative number in doItYourself
mode.

Also fix a bug which cleared entered operation numbers after an error warning
(in MathNumber, buildDigits was called on numberValue changed, which cleared
the entered value... So now called buildDigits only when really needed).