vertical_addition, don't enter digits in extra operation column in doItYourself mode
Fix to not allow entering numbers in extra column of the operation in
doItYourself mode addition, since that extra column is reserved to enter carry
and result (in case result has one more column than the operation).

Also make the main rectangles in that extra column transparent to differentiate it.

And don't add the extra column when not needed
(!items.withCarry || !level.doItYourself)