baby_wordprocessor, fix formatting issues
Fix a bug where formatting was not properly applied or saved.
It appears that when using the old method to set block tags (p/h2/h3),
it would in most cases apply only to the first block, and next blocks
had same tag as the first one but with a span style applied (so it looked ok,
but after inspecting the content returned by getFormattedText you could see
the wrong tags).

The issue appeared all the time with the following steps:
-Create a new document
-Write on first line, set it to Title (h2)
-Create a new line, type on it, set to paragraph, it looked ok but...
-Save the file
-Clear and reload the file: the 2nd line appeared with title style

So to fix this, instead of setting different tags, always keep p and apply
span styles corresponding to the various tags.