drawing_wheels, adapt pen size and offset to devicePixelRatio
Adapt pen size and offset to pixel ratio, so that available pen settings and
drawing results with same settings should be the same with different pixel
ratio screens.

Tested by doing a test file with a black shape with default QT_SCALE_FACTOR=1,
saving as SVG, then setting QT_SCALE_FACTOR=2, restarting the app and loading
the file, redoing exactly the same shape with white. The canvas position may
have a slight pixel-rounding offset (which is acceptable for our usecase target),
but other than that the shapes are exactly the same.