drawing_wheels, use style tag in SvgTank to improve compatibility with inkscape
Fix an issue I have when opening a generated svg in inkscape (v1.2), and changing
a line color, it would add a style tag with the new color but keep the old
stroke tag, making some render confused about which color to use.

Directly using the style tag instead of each separate tag fixes the issue.